Consolidate modal unit tests
