+++ packages/usa-modal/src/test/template.html
@@ -23,6 +23,16 @@
         </p>
       </div>
 
+        <label for="nestedCB" class="usa-label">Combobox label</label>
+        <div class="usa-combo-box">
+          <select class="usa-select" name="options" id="nestedCB">
+            <option value>- Select -</option>
+            <option value="value1">Option A</option>
+            <option value="value2">Option B</option>
+            <option value="value3">Option C</option>
+          </select>
+        </div>
+
       <div class="usa-modal__footer">
         <ul class="usa-button-group">
           <li class="usa-button-group__item">
